meet-asbot/pyproject.toml
[project]
name = "meet-asbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-cloud-storage>=3.6.0",
    "gunicorn>=23.0.0",
    "selenium>=4.38.0",
    "webdriver-manager>=4.0.2",
]
